Added javadoc; updated IPAddress class
