+++ kura/org.eclipse.kura.api/src/main/java/org/eclipse/kura/net/status/NetworkInterfaceState.java
@@ -13,7 +13,7 @@
 package org.eclipse.kura.net.status;
 
 /**
- * The current state of the a NetworkInterface.
+ * The current state of a network interface.
  */
 public enum NetworkInterfaceState {
